refactor calender related components
