+++ app/events/dummy/events.ts
@@ -1,0 +1,73 @@
+export const events = [
+  {
+    id: 1,
+    title: "ジャズナイト",
+    date: "2025-04-15",
+    time: "19:00 - 22:00",
+    description:
+      "地元ミュージシャンによるライブジャズパフォーマンスをお楽しみください。",
+    imageUrl: "https://picsum.photos/id/1082/800/400",
+  },
+  {
+    id: 2,
+    title: "ワインテイスティング",
+    date: "2025-04-20",
+    time: "18:00 - 21:00",
+    description:
+      "世界各国のワインを楽しみながら、ソムリエから醸造プロセスやテイスティングのコツを学びましょう。",
+    imageUrl: "https://picsum.photos/id/452/800/400",
+  },
+  {
+    id: 3,
+    title: "アートワークショップ",
+    date: "2025-04-27",
+    time: "14:00 - 17:00",
+    description:
+      "地元アーティストによる水彩画のワークショップ。初心者から上級者まで参加できます。",
+    imageUrl: "https://picsum.photos/id/49/800/400",
+  },
+  {
+    id: 4,
+    title: "オープンマイクナイト",
+    date: "2025-05-05",
+    time: "20:00 - 23:00",
+    description:
+      "詩、音楽、スタンドアップコメディなど、あなたの才能を披露する夜。参加自由です。",
+    imageUrl: "https://picsum.photos/id/96/800/400",
+  },
+  {
+    id: 5,
+    title: "クラフトビールナイト",
+    date: "2025-05-12",
+    time: "18:00 - 22:00",
+    description:
+      "地元の醸造所から厳選されたクラフトビールをお楽しみいただけます。",
+    imageUrl: "https://picsum.photos/id/24/800/400",
+  },
+  {
+    id: 6,
+    title: "コーヒーテイスティング",
+    date: "2025-05-18",
+    time: "10:00 - 12:00",
+    description: "世界各国のコーヒー豆を試飲し、その違いを学びましょう。",
+    imageUrl: "https://picsum.photos/id/766/800/400",
+  },
+  {
+    id: 7,
+    title: "ポエトリーナイト",
+    date: "2025-06-02",
+    time: "19:00 - 21:00",
+    description:
+      "詩の朗読と共有の夕べ。自作の詩を持参するか、お気に入りの詩を共有してください。",
+    imageUrl: "https://picsum.photos/id/42/800/400",
+  },
+  {
+    id: 8,
+    title: "サマーカクテルパーティー",
+    date: "2025-06-15",
+    time: "18:00 - 23:00",
+    description:
+      "夏をテーマにした特製カクテルをお楽しみください。軽食もご用意しています。",
+    imageUrl: "https://picsum.photos/id/431/800/400",
+  },
+];
