+++ app/events/types.ts
@@ -1,0 +1,8 @@
+export interface Event {
+  id: number;
+  title: string;
+  date: string;
+  time: string;
+  description: string;
+  imageUrl: string;
+}
